backend/generate-random/requirements.txt
fastapi~=0.115.6
starlette~=0.45.2
email-validator~=2.2.0
uvicorn~=0.34.0
python-dotenv~=1.0.1
pydantic~=2.10.5
pydantic-settings~=2.7.1
sqlmodel~=0.0.22
numpy~=2.2.1
alembic~=1.14.0
toml~=0.10.2
tomli~=2.2.1



# dev
flake8~=7.1.1
black~=25.1.0
isort~=6.0.0